Update for compatibility with Ui rework PR for mod loader
This is a test branch with changes to add compatibility between Hotseat and the mod loader's slightly different ui system, as well as a fix for the attack order button on the Vek turn.

It assumes base.lua has a function `Ui.onUpdate`, called each frame to allow elements to update themselves before they get relayed out and drawn.

Changed some ui elements to be translucent, to stop non-pressable elements to consume press events.

Changed menu.lua to use mod loader's functions for detecting if the escape menu is open. This solves the issue where some elements would sometimes display incorrectly for a frame.

1.2 changed how the attack order image is displayed. Instead of having an image with embedded text, the vanilla game will now display an image, and draw the text on top of it.

Added images for the attack order text, and added it to the attack order button ui element.